requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
altair==4.2.0
anyio==3.5.0
apache-airflow==2.2.4
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-sqlite==2.1.3
apispec==3.3.2
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
cached-property==1.5.2
cachelib==0.6.0
cachetools==5.0.0
cattrs==1.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==6.6.0
commonmark==0.9.1
connexion==2.13.0
croniter==1.3.4
cryptography==36.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
dnspython==2.2.1
docutils==0.16
ecdsa==0.17.0
email-validator==1.1.3
entrypoints==0.4
fastapi==0.75.1
Flask==1.1.4
Flask-AppBuilder==3.4.4
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flatbuffers==2.0
frozenlist==1.3.0
fsspec==2022.3.0
gast==0.5.3
gcsfs==2022.3.0
geographiclib==1.52
geopy==2.2.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.7.1
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-bigquery==3.0.1
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.2.3
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
graphviz==0.19.1
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.6.0
httpcore==0.14.7
httptools==0.2.0
httpx==0.22.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
importlib-resources==5.6.0
inflection==0.5.1
ipykernel==6.10.0
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
iso8601==1.0.2
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==7.2.1
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
libclang==13.0.0
lockfile==0.12.2
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.21.5
oauthlib==3.2.0
opt-einsum==3.3.0
orjson==3.6.7
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prison==0.2.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
proto-plus==1.20.3
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pydeck==0.7.1
Pygments==2.11.2
PyJWT==1.7.1
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.0.1
rsa==4.8
semver==2.13.0
Send2Trash==1.8.0
setproctitle==1.2.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
starlette==0.17.1
streamlit==1.8.1
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
ujson==4.3.0
unicodecsv==0.14.1
urllib3==1.26.9
uvicorn==0.15.0
uvloop==0.16.0
validators==0.18.2
watchdog==2.1.7
watchgod==0.8.1
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.2
Werkzeug==1.0.1
widgetsnbextension==3.6.0
wrapt==1.14.0
WTForms==2.3.3
yarl==1.7.2
zipp==3.7.0
